added the global variable & property normWavelength
This is the wavelength where the template gets normalized. If the spectrum is long enough the default is always set to 8000 angstroms, if not we find the middle of where the spectrum has values and normalize to there. We pass this wavelength value on so that the templates can be normalized there as well