first attempt at radial velocity checks for short spec
This works for spectra with wavelength coverage over all 3 shift regions, and spectra with wavelength coverage over none of the regions, have not tested the cases with some wavelength coverage yet. If there isn't full data in any of the regions, we return a nan for the rv and do not shift the spectrum.