Uses latest PyQt backend
Previously, it forced the user to use the PyQt4 backend if they were on a mac. But Python 3.6 dropped support for this backend so now it uses PyQt5 by default (if it exists) and otherwise tries to use PyQt4 (if it exists). If neither can be used, simply raises an exception.