Now using PyQt instead of Tk
This is a rewrite of the GUI frontends to PyHammer which now uses PyQt
instead of Tk. The reason being that PyQt allows for better versatility
and better features. There are also some small improvements such as
removing the necessity to supply the spectrum file type. This is NOT
fully functional and does still have bugs that need to be worked out.